Fix write serialization with inheritance
When a parent class is given, typeof(T) is not behaving the way it should and return the base class instead of the instance actual one. This change fixes an issue with emote request serialization not being possible because EmoteRequestDto is not assignable from AbstractBrowserRequestDto, the type returned by typeof(T).