Fix loading issue with skeletonNode
The problem is that the SkeletonAnimationNode is loaded with the scene (because it's a MeshNode). It waits for the UMI3DAnimatorAnimation to be loaded, but they will not arrive automatically, and if they do, they will arrive after the gltf scene has been loaded. But the client waits for the gltf scene to finish loading before displaying the scene. So it blocks.
Temporary solution found: We don't block and we put a callback to wait for the loading and replace the null value of the array. On the other hand, we send the animations ourselves.